+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -6,12 +6,11 @@ import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
 import java.net.Socket;
-import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.exception.UncheckedServletException;
 import org.apache.coyote.Processor;
-import org.apache.coyote.http11.common.RequestHeader;
-import org.apache.coyote.http11.common.RequestURI;
-import org.apache.coyote.http11.common.ResponseEntity;
+import org.apache.coyote.http11.request.RequestHeader;
+import org.apache.coyote.http11.request.RequestURI;
+import org.apache.coyote.http11.response.ResponseEntity;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
