+++ tomcat/src/main/java/org/apache/coyote/http11/request/QueryString.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.common;
+package org.apache.coyote.http11.request;
 
 import static java.util.stream.Collectors.collectingAndThen;
 import static java.util.stream.Collectors.toUnmodifiableMap;
@@ -16,7 +16,7 @@ public class QueryString {
 
     private final Map<String, String> queryStrings;
 
-    public QueryString(final Map<String, String> queryStrings) {
+    private QueryString(final Map<String, String> queryStrings) {
         this.queryStrings = queryStrings;
     }
